Extract and rename Ribbons class.
